refactored old research code and project clean up
