+++ Halo Forge Bot/Halo Forge Bot/App.xaml
@@ -1,4 +1,4 @@
-﻿<Application x:Class="Halo_Forge_Bot.App"
+﻿<Application x:Class="Halo_Forge_Bot.UI.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:Halo_Forge_Bot"
