+++ Halo Forge Bot/Halo Forge Bot/App.xaml.cs
@@ -1,12 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Configuration;
-using System.Data;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Windows;
+﻿using System.Windows;
 
-namespace Halo_Forge_Bot
+namespace Halo_Forge_Bot.UI
 {
     /// <summary>
     /// Interaction logic for App.xaml
