--- Halo Forge Bot/Halo Forge Bot/BlenderMap.cs
@@ -1,13 +1,0 @@
-using System.Collections.Generic;
-using Newtonsoft.Json;
-
-namespace Halo_Forge_Bot;
-
-public class BlenderMap
-{
-    [JsonProperty("mapId")] 
-    public int MapId { get; set; }
-    [JsonProperty("itemList")] 
-    public List<BlenderItem> ItemList { get; set; }
-}
-
