+++ Halo Forge Bot/Halo Forge Bot/MainWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="Halo_Forge_Bot.MainWindow"
+﻿<Window x:Class="Halo_Forge_Bot.UI.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
